// sdks/javascript/src/models/token.ts
export interface KnownIdToken {
  /**
   * The audience for which this token is intended.
   */
  aud: string | string[];

  /**
   * The email of the user.
   */
  email?: string;

  /**
   * The issuer identifier for the issuer of the response.
   */
  iss: string;

  /**
   * The unique human readable slug of the organization to which the user belongs.
   */
  org_handle?: string;

  /**
   * The unique identifier of the organization to which the user belongs.
   */
  org_id?: string;

  /**
   * The human readable name of the organization to which the user belongs.
   */
  org_name?: string;

  /**
   * The username the user prefers to be called.
   */
  preferred_username?: string;

  /**
   * The unique identifier of the user to whom the ID token belongs.
   */
  sub: string;

  /**
   * The tenant domain of the user.
   */
  tenant_domain?: string;
}

/**
 * Interface for ID Token payload including custom claims.
 */
export interface IdToken extends KnownIdToken {
  /**
   * Other custom claims.
   */
  [claim: string]: any;
}